Improve guest handling, session logic, and navigation
Refactored session management for reliability and better error handling. Enhanced guest user UI on Activity and Profile pages with clearer messaging and improved layout. Added direct navigation commands for Register/Login. Broadened public API route handling in AuthHandler. Updated UI text for clarity and consistency. Cleaned up unused code and improved image URL handling.

+++ MarketPrice.Ui/ViewModels/ProfileViewModel.cs
@@ -135,7 +135,7 @@ namespace MarketPrice.Ui.ViewModels
                         var isSessionEnded = await _sessionService.EndSessionAsync();
                         if (isSessionEnded)
                         {
-                            await Toast.Make("You have been logged out successfully.", ToastDuration.Long).Show();
+                            await Toast.Make("You have been logged out successfully.", ToastDuration.Short).Show();
                             await Shell.Current.GoToAsync("//Welcome");
                         }
                     }
